2005-08-14  Jb Evain  <jbevain@gmail.com>
	* Mono.Cecil/NativeType.cs
	  Mono.Cecil.Signatures/MarshalSig.cs
	  Mono.Cecil.Signatures/SignatureWriter.cs
	  Mono.Cecil.Signatures/Signature.cs:
		Work on signature writing.
	  Mono.Cecil.Implem/MethodDefinition.cs
	  Mono.Cecil.Implem/TypeDefinition.cs:
		Correct bugs in setting the hasThis field.

svn path=/trunk/cecil/; revision=48367

+++ lib/Mono.Cecil.Signatures/Signature.cs
@@ -16,7 +16,7 @@ namespace Mono.Cecil.Signatures {
 
 	internal abstract class Signature : ISignatureVisitable {
 
-		public uint CallingConvention;
+		public byte CallingConvention;
 		public uint BlobIndex;
 
 		public Signature (uint blobIndex)
